+++ example/raw/internal/transport/stream-service/health.go
@@ -3,10 +3,8 @@
 package streamservice
 
 import (
-	"errors"
 	"github.com/craftgodotdev/craftgo/pkg/server"
 	"net/http"
-	"strconv"
 
 	service "github.com/craftgodotdev/craftgo/example/raw/internal/service/stream-service"
 	types "github.com/craftgodotdev/craftgo/example/raw/internal/types/stream"
@@ -18,13 +16,9 @@ import (
 func Health(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		var req types.HealthReq
-		if _v := r.URL.Query().Get("verbose"); _v != "" {
-			_n, _err := strconv.ParseBool(_v)
-			if _err != nil {
-				server.WriteValidationError(w, r, errors.New("verbose"+": invalid bool value: "+_err.Error()))
-				return
-			}
-			req.Verbose = _n
+		_q := r.URL.Query()
+		if !server.BindValue(w, r, "verbose", "bool", _q.Get("verbose"), &req.Verbose, server.ParseBool[bool]) {
+			return
 		}
 		if err := req.Validate(); err != nil {
 			server.WriteValidationError(w, r, err)